requirements.txt
python-telegram-bot==21.11.1
pandas==2.2.3
python-dotenv==1.0.1